process_key_override: widen mod_free_replacement to uint16_t
Apple clang 21 fails the test build with:

    quantum/process_keycode/process_key_override.c:181:56: error:
      result of comparison of constant 'QK_USER' (32320) with
      expression of type 'const uint8_t' (aka 'const unsigned char')
      is always true [-Werror,-Wtautological-constant-out-of-range-compare]
      180 |     bool unregister_replacement = mod_free_replacement != KC_NO &&
      181 |                                   mod_free_replacement < SAFE_RANGE;

`clear_mods_from()` returns `uint16_t`, but the result was being
stored in a `uint8_t` local. That truncated the upper byte of the
keycode and made the `< SAFE_RANGE` (== `QK_USER` == 32320) check
tautologically true, so the "skip custom keycodes" guard never fired
for any keycode whose low byte happened to match a basic keycode.
Widening to `uint16_t` matches the function's return type and the
sibling declaration at line 349, and restores the intended check.

+++ quantum/process_keycode/process_key_override.c
@@ -175,7 +175,7 @@ const key_override_t *clear_active_override(const bool allow_reregister) {
 
     const key_override_t *const old = active_override;
 
-    const uint8_t mod_free_replacement = clear_mods_from(active_override->replacement);
+    const uint16_t mod_free_replacement = clear_mods_from(active_override->replacement);
 
     bool unregister_replacement = mod_free_replacement != KC_NO &&   // KC_NO is never registered
                                   mod_free_replacement < SAFE_RANGE; // Custom keycodes are never registered
